Упростил приведение конфигурации к значениям по-умолчанию

+++ projectTMG/main.cpp
@@ -10,7 +10,7 @@ int main()
 
 	//[ШРИФТ][ТЕКСТ][СТАНДАРТ][ДОРАБОТАТЬ]
 	unsigned int screen_x = kernel.window->getSize().x;
-	Text text(L"Привет", screen_x-200.0, 50, 20, RES_PATH); 
+	Text text(L"Привет", (float)screen_x-200, 50, 20, RES_PATH); 
 
 	//[ТЕКСТУРА][СПРАЙТ][СТАНДАРТ]
 	Sprite background(RES_PATH + "texture.png");
